+++ NEWS.md
@@ -1,5 +1,10 @@
 # graphTweets 0.4.2
 
+* Added `gt_edges_hashes_` and `gt_edges_hashes` to build networks of #hashtags co-mentions.
+* Added `%<-%` from the zeallot package to unpack the nodes and edges.
+
+# graphTweets 0.4.2
+
 Removed `splitstackshape` dependency ahead of its archival; now uses `tidyr`.
 
 # graphTweets 0.4.1